+++ tests/utils_test.lua
@@ -1,2 +1,3 @@
 require("balm/tests/u/id_generator_test")
 require("balm/tests/u/string_buffer_test")
+require("balm/tests/u/u128_test")